+++ src/main/java/com/guildsofverra/GvVersion.java
@@ -1,7 +1,7 @@
 package com.guildsofverra;
 
 public final class GvVersion {
-    public static final String CURRENT = "0.1.0-dev.6";
+    public static final String CURRENT = "0.1.0-dev.7";
 
     private GvVersion() {}
 }