+++ src/main/java/com/guildsofverra/mixin/accessor/MobEffectInstanceAccessor.java
@@ -1,0 +1,11 @@
+package com.guildsofverra.mixin.accessor;
+
+import net.minecraft.world.effect.MobEffectInstance;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(MobEffectInstance.class)
+public interface MobEffectInstanceAccessor {
+    @Accessor("duration")
+    void guildsofverra$setDuration(int duration);
+}